moved pack_bitstring, unpack_bitstring from canmatrix to utils. Sanity check on test cases passing after restructuring.
